feat: embedded dashboard, meta API, copilot sessions + hades cleanup
Phase 1: Embedded Dashboard Mode
- QylEmbedDashboard MSBuild property for embedded resource serving
- EmbeddedDashboardMiddleware with HasEmbeddedDashboard() detection
- SPA fallback: embedded mode vs dev static files

Phase 2: /api/v1/meta Capabilities
- MetaResponse split: capabilities, status, build, links, ports
- AOT-safe JSON via QylSerializerContext
- use-capabilities.ts TanStack Query hook

Phase 3: Copilot Session Resume
- CopilotSessionStore with 1h TTL + 100-session cap
- Multi-turn context enrichment in QylCopilotAdapter

Hades cleanup (-2835 lines):
- Delete dead GenAiExtractor, GenAiFields, GenAiProviders
- Delete BaggageParser, ScalarExtensions, QylAttributes
- Delete TracerProviderBuilderExtensions, IncrementalUtilities
- Consolidate servicedefaults generator
- Simplify CopilotSpanRecorder, QylTelemetry
- Add SqlOperationParser for DB instrumentation

Watchdog fix:
- Rename ToolCommandName qyl-watch → qyl-watchdog (collision with qyl.watch)
- Move core/qyl.watchdog/ → src/qyl.watchdog/ (core/ is TypeSpec-only)
- Update all CLI references, docs, launchd setup

+++ src/qyl.servicedefaults/Instrumentation/GenAi/GenAiInstrumentation.cs
@@ -371,7 +371,7 @@ public static class GenAiInstrumentation
 
         if (activity is not null && outputTokens > 0)
         {
-            activity.SetTag(GenAiAttributes.UsageOutputTokens, outputTokens);
+            activity.SetTag(GenAiAttributes.UsageOutputTokens, (int)outputTokens);
             TokenUsageHistogram.Record(outputTokens,
                 new KeyValuePair<string, object?>(GenAiAttributes.OperationName, operation),
                 new KeyValuePair<string, object?>(GenAiAttributes.ProviderName, provider),
@@ -393,7 +393,7 @@ public static class GenAiInstrumentation
     {
         if (inputTokens > 0)
         {
-            activity.SetTag(GenAiAttributes.UsageInputTokens, inputTokens);
+            activity.SetTag(GenAiAttributes.UsageInputTokens, (int)inputTokens);
             TokenUsageHistogram.Record(inputTokens,
                 new KeyValuePair<string, object?>(GenAiAttributes.OperationName, operation),
                 new KeyValuePair<string, object?>(GenAiAttributes.ProviderName, provider),
@@ -402,7 +402,7 @@ public static class GenAiInstrumentation
 
         if (outputTokens > 0)
         {
-            activity.SetTag(GenAiAttributes.UsageOutputTokens, outputTokens);
+            activity.SetTag(GenAiAttributes.UsageOutputTokens, (int)outputTokens);
             TokenUsageHistogram.Record(outputTokens,
                 new KeyValuePair<string, object?>(GenAiAttributes.OperationName, operation),
                 new KeyValuePair<string, object?>(GenAiAttributes.ProviderName, provider),
